implemented admin role

+++ app/Services/FeatureService.php
@@ -4,6 +4,7 @@ namespace App\Services;
 
 use App\Repositories\FeatureRepository;
 use Illuminate\Support\Facades\Validator;
+use Illuminate\Validation\ValidationException;
 
 class FeatureService
 {
@@ -23,9 +24,16 @@ class FeatureService
         return $this->featureRepository->store($validated);
     }
 
-    public function delete($data)   {
+    /**
+     * @throws ValidationException
+     */
+    public function delete($data){
+        if(empty($data)){
+            throw ValidationException::withMessages(['features_remove_error' => 'No chosen fields']);
+        }else{
+            return $this->featureRepository->delete($data);
+        }
 
-        return $this->featureRepository->delete($data);
     }
 
 }
